Move ApiUrlBuilder into the apiwrapper package
