+++ src/Gui/tasks/common/__init__.py
@@ -3,5 +3,4 @@ from . import utilities as utilities
 from . import dates as dates
 from .daily_recurrences_map import DailyRecurrenceMapType as DailyRecurrenceMapType
 from .daily_recurrences_map import DailyRecurrencesMapper as DailyRecurrencesMapper
-from .api_url_builder import ApiUrlBuilder as ApiUrlBuilder
 from . import macros as macros
