+++ src/Gui/tasks/services/email_verifications.py
@@ -8,7 +8,7 @@ Services for email verifications
 
 from __future__ import annotations
 from uuid import UUID
-from tasks.common.api_url_builder import ApiUrlBuilder
+from tasks.apiwrapper import ApiUrlBuilder
 from tasks.common import security
 import requests
