+++ src/Gui/tasks/services/update_password.py
@@ -4,7 +4,7 @@ import flask
 from tasks.domain import models
 from tasks.common import security
 import requests
-from tasks.common import ApiUrlBuilder
+from tasks.apiwrapper import ApiUrlBuilder
 
 class ValidatePasswordUpdateResult(Enum):
     VALID                      = 1
